backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Environment variables
python-dotenv==1.0.0

# LangChain and AI
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5

# Vector store
chromadb==0.4.22
sentence-transformers==2.3.1

# PDF processing
PyMuPDF==1.23.21
pypdf==4.0.1

# HTTP requests
httpx==0.26.0
requests==2.31.0

# Data processing
numpy==1.26.3
pandas==2.2.0

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0
